Change the structure to new flow
