+++ app/scripts/constants/ActionTypes.js
@@ -1,5 +1,6 @@
 export const API_SERVER = 'http://localhost:3000';
 //export const API_SERVER = 'http://52.74.99.100';
+//export const API_SERVER = 'http://54.254.159.95';
 
 export const LOGIN = 'LOGIN';
 export const LOGIN_SUCCESS = 'LOGIN_SUCCESS';
