+++ app/styles/app.css
@@ -36,10 +36,6 @@ pre[class*="language-"] {
   padding: 16px;
   color: white;
   text-align: center;
-  button{
-    float:right;
-    padding: 10px 20px;
-  }
 }
 
 .screen {
@@ -49,6 +45,23 @@ pre[class*="language-"] {
   overflow: auto;
   width: 100%;
   height: 100%;
+  button.showMap{
+    position: absolute!important;
+    top: 0;
+    right: 0;
+    padding: 10px 20px;
+    z-index:9999;
+    margin: 15px 10px;
+  }
+
+  button.giveUp{
+    position: absolute!important;
+    top: 65px;
+    right: 0;
+    padding: 10px 20px;
+    z-index:9999;
+    margin: 15px 10px;
+  }
 }
 
 .login {
@@ -88,13 +101,17 @@ pre[class*="language-"] {
   }
 }
 
-.activity-card {
+.activity-card-container {
   position: relative;
-  max-width: 592px;
+  /*max-width: 1000px;*/
   margin: 30px auto 80px;
 
-  p {
-    margin-top: 0;
+  .activity-card {
+    float: left;
+    max-width: 495px;
+    p {
+      margin-top: 0;
+    }
   }
 }
